fix: build fs/zfs on FreeBSD prior to 13
d6dab2c broke fs/zfs on FreeBSD versions prior to 13. Some of the
ZFS-related symbols included in that commit are not in FreeBSD's ZFS
implementation (but were found in OpenZFS). FreeBSD did not adopt
OpenZFS until FreeBSD 13.

This commit enables building of fs/zfs on FreeBSD both with and without
OpenZFS.